Factor out non-Windows code
